requirements.txt
python-dotenv
requests
openai
pyaudio